Rename order app group method names and route
